log: Fix issue if server crashes while flushing data to log
In such a case, the log may be truncated without adding the Footer to
the end. The only way around this would be to copy the part of the file
to be kept and write the new part and the Footer to the end and then
move it into place.

To make things more better, move the Footer out to its own file which
will be updated after the log file. This will allow the log file update
to fail while the old header file remains (somewhat) accurate for the
state of the previous log file.

If replaying the WAL, then the WAL is not cleared until the log file AND
header are successfully re-written.