storage: Slight performance improvements
For the write-ahead log, avoid reopening the file for every call to
::extend. Instead, delay closing it with a new LazilyClosed type.

Secondly, for the write-ahead, create a formal asynchronous task to
manage the delay and auto-replay of the log to the persistence backend.
This ends up eliminating a re-entrant issue where the replay method
might have been called concurrently with the previous pattern.